+++ tools/json-generator/src/app/get-codewars-challenges.ts
