+++ tools/json-generator/src/app/get-codewars-user.ts
