+++ tools/json-generator/src/app/get-git-info.ts
